Unquantize Points during presimplify
Simplifying a topology requires the removal of the transform (if any).
This is done it the presimplify() function.

presimplify() only does that on arcs, not points.
If the topology is quantized again later, points will be quantized
twice, yielding invalid values.

With this commit, presimplify() also remove the transform of Point &
MultiPoint objects.